Update first in ship chain instead in all cases
This handles not only notmrla updates but also if youre updating a staged update with a reshippedFromId, it'll update the root parent instead. All updates should happen to the root ship as it is the source of truth. After this commit, I need to change draft ships to pull their titles etc (all info) from the root ship